Provide separate accessors for DocContent and DocContentOwner
Previously, in order to access the `DocContent` for say a block tag it
was necessary to get the DocContentOwner` and then access its
`docContent`. That causes problems when attempting to mutate a
non-existent block tag.

The current behavior of `appendDocumentation(String, String?)` is that
if a `tagSection` is provided and it does not exist then appending will
create it. Replicating that in terms of the `DocContentOwner`will
require that a `DocContentOwner` is created for the non-existent block
tag that will create the block tag when content is appended to it.

That means every time that the `DocContentOwner` is requested for a
block that that it will need to be created if it does not exist. If it
is just being retrieved to access its `DocContent` then that is a lot
of wasted effort and space.

This change avoids that by providing separate accessors for the
`DocContent` and the `DocContentOwner` and using the former where
possible. Follow-up changes will switch all the `DocContentOwner?`
accessors to `DocContentOwner` that will handle the creation of a
block tag where necessary.

Bug: 450228132
Test: ./gradlew
Change-Id: I01e224591c5b7b4228076ad6a4bc7a20c2fc450a