Replace ParameterDefaultValueFactory with hasDefaultValue boolean
All `ParameterDefaultValue` did was act as a wrapper for a boolean representing whether a parameter has a default value. This removes it and just uses a boolean directly.

Now that `PsiParameterDefaultValue` doesn't require a `PsiParameterItem` to be created before computing whether the parameter has a default value, all parameter initialization can get rid of the `defaultValueFactory` and instead directly compute whether a parameter has a default value.

Bug: 454079088
Test: ./gradlew test

Change-Id: If29dabc1cd8c2f79177562eae7d67de4becf1770