Add tests to show behavior when appending to overriding method
This change adds tests for `ItemDocumentation.appendDocumentation(...)`
and `DocContentOwner.append(...)` appending to an overriding method
without any existing comment. They show that the latter does not add an
`{@inheritDoc}`. That is needed as an overriding method with no comment
will automatically inherit documentation from the overridden method but
adding a comment (even an empty one) can change that behavior.

It also shows that the order in which content is appended to the main
description and to a non-existing block tag (forcing it to be added)
affects the results. That, and other issues will be addressed in
b/454257440.

A follow-up change will fix `DocContentOwner.append(...)` to behave the
same as `appendDocumentation(...)` for now to minimize changes in the
stubs.

Bug: 450228132
Test: ./gradlew
Change-Id: I7f7152c75093896355f84422fd114b9fbc82f2cf