Make DocContentOwner.append(...) add {@inheritDoc} when needed
Previously, `DocContentOwner.append(...)` did not behave the same way
as `ItemDocumentation.appendDocumentation(...)` when it was appending
content to a `DocComment` that did not exist in the sources and which
was attached to an overriding method. It failed to insert
`{@inheritDoc}`.

This change makes it behave in the same way. It is not clear that is
the correct behavior but it avoids changes that could cause unexpected
issues in the generated HTML. Bug 454257440 has been opened to
ensure correct `{@inheritDoc}` behavior.

Bug: 450228132
Test: ./gradlew
Change-Id: I93bf35503a56bcf6b63c2286e2988081c2ec2793