[Readability Refactor] Simplify Api Levels Generation patterns logic
The logic for determining the api signature file's patterns and files is
a bit hard to read. E.g. if you have to read a big <if> statement to
undetstand what each value of the refactored triple represents.

This CL reorders the logic so you can immediately see what each value
represents without having to read the details of other values

The error state is now clearly laid out at the top of the method
execution which also simplifies the logic

This CL:
* Removes Triple value if assignment
* Renames variables to make it clear what kind of file is being
  extracted (i.e. adding "api" to variable names)

Bug: 450531827
Test: ./gradlew
Change-Id: I8e55c0cc658a22e8b4e0482b84effa690af533bb