[Refactor Prep/Readability] Extract logic for CodebaseSdkVersion
This CL introduces a new method to encapsulate all the decisions that go
into deciding the api version for the current codebase, including
whether it should be included in the api history.

Why did I do this:
* Easier to read
* The follow up CLs will directly tackle how we go about simplifying the
  amount of options Metalava relies on to determine the api version of
sources. This provides a clear home for such logic which will under go
many incremental changes
* We plan to remove --current-version eventually, I want to encapsulate
  where it is being used temporarily so it's easier to see how to avoid
breaking changes while we work on removing it

Bug: 450531827
Test: ./gradlew

Change-Id: Ib3b66ecdd8cc19dd33f51173385d06043f52489a