Make append(DocContent) test more realistic
Currently, the only place where documentation from one part of the
source is appended to documentation for another part is when
documentation from annotations is copied to the annotated items. In
those cases the documentation is retrieved from block tags on the
annotations, e.g. `@memberDoc`, `@classDoc`, etc..

That documentation needs to be processed specially in two ways:
1. Leading whitespace needs to be removed.
2. It needs to be fully qualified relative to the annotation from which
   it was taken.

Fixing that generally will require a lot of work. So, in the meantime
we need a way to treat that documentation specially. This change
prepares for that by tweaking the test to make it more realistic by
getting the content from the `@memberDoc` block tag. It also adds a
`{@link}` tag to a class to test how they are fully qualified.

Bug: 450228132
Test: ./gradlew
Change-Id: Iba13123628bde8ded05aee88cf054568f54e8a20