Fix bug in PsiItemDocumentation.findMainDocumentation()
Previously, it assumed that the `text` that was used to construct
`comment` was the same length as `comment.text`, or at least no bigger.
While that was mostly true it seems as though there is at least one
case when that is not true and so the call to `substring(...)` fails
with a `StringIndexOutOfBoundsException`. Some future changes to switch
to the `DocComment` model appear to cause that to happen.

It is not clear why this happens but it is not important as the
`findMainDocumentation()` method will be removed soon so it will become
moot.

In the mean time this change uses `comment.text.length` instead of
`text.length` which fixes the problem.

Bug: 450228132
Test: ./gradlew
Change-Id: Ic277dfdb83fbf173eede0abbb7849b9ad0cb8f81