Use containsWord("null") in AndroidApiChecks.checkVariable(...)
Previously, it used a regular expression to search for `null` as a
word. This switches it to use `containsWord("null")` as that is faster,
and consistent with how other checks are performed.

The null word check was the first part of an `&&` expression. As it is
quite expensive, certainly compared to the second part this change also
switches them around to avoid doing the expensive part unless
necessary.

Bug: 450228132
Test: ./gradlew
Change-Id: I57fb10712face9fe9479643139f3fb297b8748bd