Add support for applying custom predicates to DocContent
Currently, various checks apply custom predicates to the textual form
of the documentation. This change adds the ability to do the same to
`DocContent`.

These are needed by a number of the following changes so they are in
their own change to avoid coupling it with a specific change that uses
it. That provides more flexibility when managing the stack of changes.

Bug: 450228132
Test: ./gradlew
Change-Id: I4836e10963e1e003c4a0ccc7be66efe5636787be