Mark package as compatibility suppressed
Before this change, packages would get marked as compatibility
suppressed if their child classes were all experimental. However, this
created problems when a parent package has an experimental child
class, and also child packages with non-experimental classes. In this
case, the child package would also get marked as experimental because
the `isCompatibilitySuppressed()` method of the descendant classes
scans all parents. An example can be seen here:
https://android-review.git.corp.google.com/c/platform/frameworks/support/+/3814424/3
in `compose/animation/animation-graphics/api/1.10.0-beta01.txt.

This CL changes the process and only marks a package as compatibility
suppressed in a signature file if all child classes and packages are
experimental.

Bug: 404795417
Test: `./gradlew`
Change-Id: Ia7051928911e3751f5698a63f598c676dffaefde