Stop tracking compiler-generated default overloads for reified functions
Previously, tracking was added for the compiler-generated `$default` version of Kotlin functions with optional parameters (aosp/3788272). This included reified inline functions. However, we don't track bytecode versions of reified inline functions generally because they cannot be used from bytecode without a runtime error (reified inline functions can only be used from Kotlin source, and will be inlined in the bytecode anywhere they are used).

This updates how kotlin metadata is found for `$default` functions so that it will be possible to tell from the metadata if `$default` functions are reified inline.

Bug: 454620291
Test: Added test to verify that for reified functions with optional parameters only the kotlin source signature is tracked. Without this change the `$default` overloads were also tracked.

Change-Id: I9b0ba6625914189cd0be159ac6208990058b3622