[Add test/prep] Add test for --api-versions-for-sources REL behaviour
currently --api-versions-for-sources has no effect when the codename is
rel. Adding a test case to show case this.

Why am i doing this:
In the follow up CLs this behaviour will change where the
codebaseSdkValue will be explicity set to --api-versions-for-sources no
matter what if it is provided. Temporarily, in the short term we will fall
back to the current behaviour of --current-api-versions (which will
eventually be removed)

Bug: 454053322
Test: CorrectApiLevelForNonReleaseTest.kt

Change-Id: I672afff53dc8a7584d6a59082f6ee50110fae278