CodebaseSdkValue = --api-version-for-sources when its is set
This CL:
* Updates CodebaseSdkValue logic to always use
  --api-versions-for-sources if it is set to a valid value
* Update test case to now assume --api-versions-for-sources is used
  regardless of codename
* Removed default value of --api-versions-for-sources

Context and future work:
Eventually --api-versions-for-sources will be the only argument used to
set the api level of sources. As a fallback, if it is not provided we
continue to use the current --current-versions behaviour.

Follow up CLs plan to eventually phase out --current-versions all
together. When that happens, the lack of --api-versions-for-sources will
imply that sources should not be included in the api history.

Bug: 454053322
Test: ./gradlew
Change-Id: Ib45c92240f643bef57bc6444ff4a65756cdba912