Add an option to run metalava build with remote maven repos
We need to use a prebuilts repo when running CI because maven central is throttling requests from google. However, the androidx prebuilts repo is not part of the manifest for android, so it currently isn't possible to build metalava with gradle in main. This adds an option to use remote maven repos with the `DOWNLOAD_DEPENDENCIES` env var.

Bug: 454872806
Test: With the prebuilts repo present, `./gradlew publish` runs successfully. Without the prebuilts repo, `./gradlew publish` fails, but `DOWNLOAD_DEPENDENCIES=true ./gradlew publish` passes.
Change-Id: Ic859afa933f7d11510a237685a62d7906ac2b5bd