Added test for adding final to experimental method
Before this CL, converting a method to final would not be flagged if
that method was marked as experimental/compatibility suppressed.
However, converting an experimental method to final can create breaking
changes for clients who have to implement that method in order to use a
non-experimental abstract class.

This commit just adds a test to demonstrate the issue, the followup
commit will create the fix.

Bug: 454033024
Test: `./gradlew`
Change-Id: I4efff9d6667b9d139a5ab86bb183342d351b71b4