Add addLeadingSpaceIfNeeded to JavadocContentPrinter.print(...)
Previously, it was the caller's responsibility to determine if a space
needed to be added before the content. Sometimes that was fixed and
sometimes it depended on whether the content started with a newline
character. This change adds support in the `JavadocContentPrinter` to
check the latter and possibly allow for future optimizations.

Preparation for changes that will allow `TagType` specific formatting
of the tag content and so increase the number of callers of the
`print(...)` method.

Note: Previously, block tags always added a space before the content
(if there was any) because all leading whitespace (including newlines)
had been removed from `description`. However, while it will have no
effect this change uses `addLeadingSpaceIfNeeded = true` when printing
block tags to be consistent with inline tags.

Bug: 447588621
Test: ./gradlew
Change-Id: I27ec935791011f741ae29c0933fb41c354144e35