Compat check for making experimental method final
Before this CL, converting a method to final would not be flagged if
that method was marked as experimental/compatibility suppressed.
However, converting an experimental method to final can create breaking
changes for clients who have to implement that method in order to use a
non-experimental abstract class.

This commit implements the fix.

Bug: 454033024
Test: `./gradlew`
Change-Id: Iad17c210b4533d7fd8ab18435c7b1504db583bc2