Add test for ChangedType experimental compat
As of now, if an experimental method changes its return type at all, no
compatibility issues are raised. However, we should be raising an error
for changed return type on an experimentally marked method if that
method is part of an abstract class or interface and a client is forced
to implement it.

This commit just adds a test demonstrating the issue, the follow-up
commit will implement the fix.

Bug: 454029984
Test: `./gradlew`
Change-Id: I5310affed5984a655fb840e53e4c26418a8c54c6