Raise compat error on ChangedType for method
As of now, if an experimental method changes its return type at all, no
compatibility issues are raised. However, we should be raising an error
for changed return type on an experimentally marked method if that
method is part of an abstract class or interface and a client is forced
to implement it.

This commit detects and raises a compatibility error.

Bug: 454029984
Test: `./gradlew`
Change-Id: I51d76ee01c1e77e765be1884eafc5662552a18e1