Always invalidate documentation.text after parsing DocComment
It is necessary for `AbstractItemDocumentation` to keep `text` and
`docComment` in sync with each other so that code which uses one or the
other have a consistent view of the comment. Previously, it was assumed
that `DocComment` was identical to the text from which it was parsed.
However, that is not true. `DocComment` normalizes whitespace, resolves
`@throws` which change its textual form.

This change addresses that by invalidating `text` after parsing
`docComment` forcing it to be regenerated the next time it is accessed.
That results in a few test changes that needed fixing.

Bug: 447588621
Test: ./gradlew
Change-Id: Ic7b9303d3e275d5827df44913fe711b8c5a009b6