Simplify PsiItemDocumentation by removing samePackage(...)
Previously, `PsiItemDocumentation` used `samePackage(...)` to simplify
and reduce the size of resolved references but it was disabled because
it caused problems with `doclava`. The intent was to enable it in
future when `doclava` was no longer used. However, this content is
going to be removed long before we switch away from `doclava` so there
is no point in keeping as it just complicates the already hard to
understand code.

Therefore, this change removes `samePackage(...)`. It will be easy to
add similar logic for references in the new model as there will be just
one place that will need modification.

Bug: 447588621
Test: ./gradlew
Change-Id: I33ad5f3acd007c915c4313810debabf21c224204