Propagate showability annotations from properties to accessors
The PublishedApi annotation means an API needs to be tracked for compatibility because even though the internal API can't be used directly by an external client, it can be used by a non-internal inline function which can then be used by an external client, which means the reference to the PublishedApi will end up inlined in the bytecode of the external client. For a regular property, it is a reference to the accessor that will be inlined, so a PublishedApi property should have its accessors tracked for binary compatibility.

This required working around a bug in UAST where these accessors end up having mangled names even though they should not (b/457844210, https://youtrack.jetbrains.com/issue/KT-82182).

Test: fixes behavior in ShowAnnotationTest, androidx changes: aosp/3837158
Bug: 450844478
Change-Id: I69480d9e62f0ea19d76fead54bd147c9dc2bea80