Add tests demonstrating property annotation behavior
This shows how experimental and show annotations on properties are currently treated with respect to the associated accessor methods and backing fields.

Experimental annotations are propagated to both accessors and fields, which is the desired behavior because the kotlin compiler will treat references to both as experimental.

Show annotations (i.e. PublishedApi) are currently propagated to fields but not accessors. They should instead be propagated to accessors when they exist (for regular properties, since this is how the properties will be references), and only propagated to fields when the backing field is how the property will be referenced (consts and JvmStatic properties).

Bug: 450844478
Bug: 450916834
Test: adds tests
Change-Id: I8cfa184d986a0a104f3122dfa37fcbf3de374ab6