Add TurbineItemDocumentation to compute FileLocation correctly
Previously, Turbine did not compute the `FileLocation` for comments
which meant that tests which checked issues reported in the javadoc
failed on Turbine. This change adds a `TurbineItemDocumentation`
instance which computes it correctly and fixes those tests.

Preparation for a follow-up that adds more such tests that would
require baselining the tests for Turbine without this.

Bug: 447588621
Test: ./gradlew
Change-Id: If84a0252c2f0df1632b7e22fe7cb68b81c5db774