Report TurbineError diagnostics
Previously, if turbine threw a `TurbineError` it would abort all of
Metalava's processing. This change catches it and reports all the
diagnostics and carries on. That has the potential to leave the
`Codebase` in an invalid state but that is true of all such failures
whether in Turbine or Psi.

Bug: 447588621
Test: ./gradlew
Change-Id: I1c64f9e2e52b00dd0afb4886cb02bd5aa31e4c40