Add LAX_PARSER Capability to model providers
Previously, there was no way for a test to differentiate between Psi
and Turbine's main difference which is that the former is a lax parser
(can continue after parse errors) while Turbine is very strict. This
change adds a `LAX_PARSER` capability and uses that in the
`CommonDuplicateClassItemTest` to only check the `Codebase` structure
for providers that have a `LAX_PARSER` capability, i.e. only on Psi
and not Turbine.

Bug: 447588621
Test: ./gradlew
Change-Id: I7a24af79782c87c6afe01e1f41fc33a42ee26b12