Unwrap repeated annotations in Turbine
Previously, Turbine would add the whole container for a repeated
annotation to an item. That differed from the Psi behavior and is
unhelpful as the container is really an implementation issue to
workaround limitations in the `.class` byte code format and the JVM.
The source representation uses the individual annotations.

This change makes Turbine unwrap the container like Psi does, fixing
the test that was added in the previous change.

Bug: 354633349
Test: ./gradlew
Change-Id: I43beefa9c9d74a760f2c0dbac1623c71a5b9124d