Track exhaustivity for sealed classes
Instances of sealed classes can be matched using `when` statements. If
all the subclasses of a sealed class are available to API consumers,
then new subclasses can't be added to the sealed class because doing
so would be a breaking change (clients' `when` statements would no
longer be exhaustive). In this case, we label the sealed class as
exhaustive. For more details, see b/447143803

This commit updates signature file writing, the followup commit will
include the associated compatibility check.

Bug: 447143803
Test: `./gradlew`
Change-Id: Ic4af999934b5a8d71cc75ba34ef3fcd195d5a022