Make type aliases ClassItems
Having a TypeAliasItem separate from ClassItem makes sense in many ways because typealiases are very different than regular classes. However, with KMP, it is possib
le to have an `expect class` have an `actual typealias`, and modeling this case will be much simpler if typealiases are `ClassItem`s.

One consequence of this is that some checks for classes are now applied to typealiases as well (see aosp/3823463), which seems good.

Bug: 458735166
Test: ./gradlew test

Change-Id: Ic55c220c1a3774afb9951ac05ac60fd22beab478