Refactor type nullability compatibility check
Previously, `compareItemNullability` could only compare the nullability of the `type()` of an `Item`, but an `Item` may have other associated types.

This splits the type comparison into another function where both the types and associated items can be specified, which will be helpful in followups for comparing nullability of aliased types when they are modeled as classes, and for comparing nullabiity of property receivers, which isn't currently done.

Bug: 458735166
Bug: 300126192
Test: ./gradlew test
Change-Id: I3343ebef2651a812ea058c17c16e811015393db0